app/agents/resume_tailor.py
"""Resume Tailoring Agent implementation using Google ADK."""

import json
import os
from typing import Any, Dict, List, Optional, Union

from loguru import logger
from pydantic import BaseModel, Field

from app.models.job_model import JobPosting
from app.models.resume_model import Resume
# from backend.app.agents.adk_integration import create_agent, execute_agent # To-Do: Restore or reimplement ADK integration logic
# To-Do: Define or import create_agent and execute_agent if they are to be used directly
create_agent = None # Placeholder
execute_agent = None # Placeholder

from google.adk.tools import Tool, ToolConfig, ToolSpec

# Define tool specs for the resume tailoring agent
HIGHLIGHT_SKILLS_TOOL = ToolSpec(
    name="highlight_skills",
    description="Identify and highlight skills in the resume that match the job requirements",
    parameters={
        "matched_skills": {
            "type": "array",
            "description": "List of skills found in both resume and job description",
            "items": {"type": "string"}
        }
    },
    response_format={"type": "object"}
)

REORDER_EXPERIENCE_TOOL = ToolSpec(
    name="reorder_experience",
    description="Reorder work experience to prioritize relevant experience",
    parameters={
        "experience_order": {
            "type": "array",
            "description": "List of work experience IDs in the order they should appear",
            "items": {"type": "string"}
        },
        "relevance_scores": {
            "type": "object",
            "description": "Dictionary of experience ID to relevance score",
            "additionalProperties": {"type": "number"}
        }
    },
    response_format={"type": "object"}
)

ENHANCE_BULLET_POINTS_TOOL = ToolSpec(
    name="enhance_bullet_points",
    description="Enhance resume bullet points to better match job requirements",
    parameters={
        "enhanced_bullets": {
            "type": "object",
            "description": "Dictionary of experience ID to list of enhanced bullet points",
            "additionalProperties": {
                "type": "array",
                "items": {"type": "string"}
            }
        }
    },
    response_format={"type": "object"}
)


class ResumeTailoringAgent:
    """Agent for tailoring resumes to job descriptions using Google ADK."""
    
    def __init__(self):
        """Initialize the resume tailoring agent."""
        # To-Do: Implement agent initialization
        pass
    
    def _load_system_prompt(self) -> str:
        """Load the system prompt for the resume tailoring agent."""
        # To-Do: Implement system prompt loading
        pass
    
    def _create_tools(self) -> List[Tool]:
        """Create the tools for the resume tailoring agent."""
        
        # Implement tool functions
        def highlight_skills(matched_skills: List[str]) -> Dict[str, Any]:
            """Highlight skills tool implementation."""
            # To-Do: Implement highlight_skills tool logic
            pass
        
        def reorder_experience(experience_order: List[str], relevance_scores: Dict[str, float]) -> Dict[str, Any]:
            """Reorder experience tool implementation."""
            # To-Do: Implement reorder_experience tool logic
            pass
        
        def enhance_bullet_points(enhanced_bullets: Dict[str, List[str]]) -> Dict[str, Any]:
            """Enhance bullet points tool implementation."""
            # To-Do: Implement enhance_bullet_points tool logic
            pass
        
        # To-Do: Implement tool creation and configuration
        pass
    
    def tailor_resume(self, resume: Resume, job: JobPosting, format_type: str = "text") -> Dict[str, Any]:
        """Tailor a resume to a job description using structured data.
        
        Args:
            resume: The resume to tailor.
            job: The job description to tailor the resume to.
            format_type: The format of the output (text, html, docx, pdf).
            
        Returns:
            A dictionary with the tailored resume and additional metadata.
        """
        # To-Do: Implement resume tailoring logic
        pass
        
    def generate_tailored_resume(self, resume_data: Dict[str, Any], job_data: Dict[str, Any], format_type: str = "text") -> Dict[str, Any]:
        """Generate a tailored resume from arbitrary data formats.
        
        This method is more flexible than tailor_resume as it accepts any format of resume and job data,
        making it suitable for use with data coming from the frontend or API.
        
        Args:
            resume_data: The resume data in any format.
            job_data: The job description data in any format.
            format_type: The format of the output (text, html, docx, pdf).
            
        Returns:
            A dictionary with the tailored resume content and improvements made.
        """
        # To-Do: Implement tailored resume generation from arbitrary data
        pass
    
    def _extract_improvements(self, response: Dict[str, Any]) -> List[str]:
        """Extract improvements made from the agent response.
        
        Args:
            response: The agent response.
            
        Returns:
            A list of improvements made.
        """
        # To-Do: Implement improvements extraction from response
        pass
